Added a method to make delivery to the local mailbox conditional on the presence of a certain value in some other attribute.  Used to implement mailForwardingAddress both in addition to normal delivery and excluding normal delivery, selectable entry by entry.  The model is mailDeliveryOption in Netscape MS.  The implementation aims to become more general, though.  Affects "search-with-filter", any entry can potentially use a parameter, introduced with "param=".
Optimize the case where we have to copy the message to an address that
is served by the directory.  Formerly, we would have the MTA deal with
it and invoke mail500 again later.  This has necessitated loading the
list of domains that are solved by us with "domain".  A new definition,
"host", takes the role of the old "domain" that was the FQDN of our
host for routing loop avoidance.