ITS#2717, use default SASL_MECH if none was provided.
Note that the mech stored in the void *defaults structure is useless
because we can't look inside the structure, and the mech list is needed
as input to the SASL library, before any callbacks are called. It seems
the mech doesn't belong in the *defaults struct, and/or it should be
visible and not opaque.