Return value instead of key for getCurrentString et al in Cursor
Affects: getCurrentString(), getCurrentStringUnsafe(), getCurrentBinary(), getCurrentBinaryUnsafe(), getCurrentNumber(), and getCurrentBoolean()
When using a Cursor, the key is always available as the return value from goToNext() et al, and some (including myself) find using a callback to access the value is unnecessary and slow. This change returns the value directly, instead of the key. The callback interface is still enabled if a function is passed, however, so this shouldn't affect existing code.