Issue#8769 - Fix oid search extension syntax in documentation
The option string, e.g. '<oid>=:dn:' is parsed like a LDIF entry starting from
the '=' and replacing the '=' with a dummy variable 'x'.  In this case, said
string is 'x:dn:', so the resulting effective value is 'dn:'.  This also implies
that base64 values have to be passed in the form '<oid>=::<b64value>'.